# Optimization based on profiling for forward