--- a/csrc/flash_attn_rocm/src/flash_fwd_runner_kloop_hdim128_fp16_noncausal_gfx90a.cpp
+++ b/csrc/flash_attn_rocm/src/flash_fwd_runner_kloop_hdim128_fp16_noncausal_gfx90a.cpp
@@ -26,10 +26,23 @@
 namespace fwd_device_gemm {
 // hdim 128, fp16, non-causal
 template <>
-void FlashFwdRunner::Run<false, 128, device_gemm_trait::Float16, false>() {
+void FlashFwdRunner::Run<false, 128, device_gemm_trait::Float16, false, true>() {
   BOOL_SWITCH(is_deterministic_, kIsDeterministic, [&] {
     this->template run_<DeviceGemmKLoopHeadDim128,
                   device_gemm_trait::Float16, 
+                  device_gemm_trait::kGemmSpecPadding,
+                  device_gemm_trait::kMaskingSpecDefault,
+                  kIsDeterministic>();
+  });
+} // FlashFwdRunner::Run()
+
+// hdim 128, fp16, non-causal
+template <>
+void FlashFwdRunner::Run<false, 128, device_gemm_trait::Float16, false, false>() {
+  BOOL_SWITCH(is_deterministic_, kIsDeterministic, [&] {
+    this->template run_<DeviceGemmKLoopHeadDim128,
+                  device_gemm_trait::Float16, 
+                  device_gemm_trait::kGemmSpecPadding,
                   device_gemm_trait::kMaskingSpecDefault,
                   kIsDeterministic>();
   });