--- a/csrc/flash_attn_rocm/src/flash_fwd_runner_kloop_hdim128_fp16_noncausal_gfx90a.cpp
+++ b/csrc/flash_attn_rocm/src/flash_fwd_runner_kloop_hdim128_fp16_noncausal_gfx90a.cpp
@@ -24,7 +24,7 @@
 #include "flash_fwd_runner_gfx90a.h"
 
 namespace fwd_device_gemm {
-// hdim 128, fp16, non-causal
+// hdim 128, fp16, non-causal, MNKO-padding
 template <>
 void FlashFwdRunner::Run<false, 128, device_gemm_trait::Float16, false, true>(bool is_dropout) {
   BOOL_SWITCH(is_deterministic_, kIsDeterministic, [&] {
@@ -36,13 +36,13 @@ void FlashFwdRunner::Run<false, 128, device_gemm_trait::Float16, false, true>(bo
   });
 } // FlashFwdRunner::Run()
 
-// hdim 128, fp16, non-causal
+// hdim 128, fp16, non-causal, non-padding
 template <>
 void FlashFwdRunner::Run<false, 128, device_gemm_trait::Float16, false, false>(bool is_dropout) {
   BOOL_SWITCH(is_deterministic_, kIsDeterministic, [&] {
     this->template run_<DeviceGemmKLoopHeadDim128,
                   device_gemm_trait::Float16, 
-                  device_gemm_trait::kGemmSpecPadding,
+                  device_gemm_trait::kGemmSpecDefault,
                   device_gemm_trait::kMaskingSpecDefault,
                   kIsDeterministic>();
   });