--- a/csrc/flash_attn_rocm/src/flash_fwd_runner_kloop_hdim32_bf16_causal_gfx90a.cpp
+++ b/csrc/flash_attn_rocm/src/flash_fwd_runner_kloop_hdim32_bf16_causal_gfx90a.cpp
@@ -24,12 +24,25 @@
 #include "flash_fwd_runner_gfx90a.h"
 
 namespace fwd_device_gemm {
-// hdim 32, bf16, causal
+// hdim 32, bf16, causal, MNKO-padding
 template <>
-void FlashFwdRunner::Run<false, 32, device_gemm_trait::BFloat16, true>() {
+void FlashFwdRunner::Run<false, 32, device_gemm_trait::BFloat16, true, true>() {
   BOOL_SWITCH(is_deterministic_, kIsDeterministic, [&] {
     this->template run_<DeviceGemmKLoopHeadDim32,
                   device_gemm_trait::BFloat16, 
+                  device_gemm_trait::kGemmSpecPadding,
+                  device_gemm_trait::kMaskingSpecCausal,
+                  kIsDeterministic>();
+  });
+} // FlashFwdRunner::Run()
+
+// hdim 32, bf16, causal, MNKO-padding
+template <>
+void FlashFwdRunner::Run<false, 32, device_gemm_trait::BFloat16, true, false>() {
+  BOOL_SWITCH(is_deterministic_, kIsDeterministic, [&] {
+    this->template run_<DeviceGemmKLoopHeadDim32,
+                  device_gemm_trait::BFloat16, 
+                  device_gemm_trait::kGemmSpecPadding,
                   device_gemm_trait::kMaskingSpecCausal,
                   kIsDeterministic>();
   });