--- a/csrc/flash_attn_rocm/src/flash_fwd_runner_kloop_hdim32_bf16_noncausal_gfx90a.cpp
+++ b/csrc/flash_attn_rocm/src/flash_fwd_runner_kloop_hdim32_bf16_noncausal_gfx90a.cpp
@@ -26,7 +26,7 @@
 namespace fwd_device_gemm {
 // hdim 32, bf16, non-causal, MNKO-padding
 template <>
-void FlashFwdRunner::Run<false, 32, device_gemm_trait::BFloat16, false, true>() {
+void FlashFwdRunner::Run<false, 32, device_gemm_trait::BFloat16, false, true>(bool is_dropout) {
   BOOL_SWITCH(is_deterministic_, kIsDeterministic, [&] {
     this->template run_<DeviceGemmKLoopHeadDim32,
                   device_gemm_trait::BFloat16, 
@@ -38,7 +38,7 @@ void FlashFwdRunner::Run<false, 32, device_gemm_trait::BFloat16, false, true>() 
 
 // hdim 32, bf16, non-causal, MNKO-padding
 template <>
-void FlashFwdRunner::Run<false, 32, device_gemm_trait::BFloat16, false, false>() {
+void FlashFwdRunner::Run<false, 32, device_gemm_trait::BFloat16, false, false>(bool is_dropout) {
   BOOL_SWITCH(is_deterministic_, kIsDeterministic, [&] {
     this->template run_<DeviceGemmKLoopHeadDim32,
                   device_gemm_trait::BFloat16,