--- a/csrc/flash_attn_rocm/src/flash_fwd_runner_kloop_hdim32_fp16_causal_gfx90a.cpp
+++ b/csrc/flash_attn_rocm/src/flash_fwd_runner_kloop_hdim32_fp16_causal_gfx90a.cpp
@@ -26,7 +26,7 @@
 namespace fwd_device_gemm {
 // hdim 32, fp16, causal, MNKO-padding
 template <>
-void FlashFwdRunner::Run<false, 32, device_gemm_trait::Float16, true, true>() {
+void FlashFwdRunner::Run<false, 32, device_gemm_trait::Float16, true, true>(bool is_dropout) {
   BOOL_SWITCH(is_deterministic_, kIsDeterministic, [&] {
     this->template run_<DeviceGemmKLoopHeadDim32,
                   device_gemm_trait::Float16, 
@@ -38,7 +38,7 @@ void FlashFwdRunner::Run<false, 32, device_gemm_trait::Float16, true, true>() {
 
 // hdim 32, fp16, causal, MNKO-padding
 template <>
-void FlashFwdRunner::Run<false, 32, device_gemm_trait::Float16, true, false>() {
+void FlashFwdRunner::Run<false, 32, device_gemm_trait::Float16, true, false>(bool is_dropout) {
   BOOL_SWITCH(is_deterministic_, kIsDeterministic, [&] {
     this->template run_<DeviceGemmKLoopHeadDim32,
                   device_gemm_trait::Float16,